requirements.txt
# Google AI and Generative AI
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-cloud-texttospeech>=2.27.0

# Async and Event Loop
asyncio
aiohttp>=3.8.0
uvloop>=0.17.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Communication and Messaging
websockets>=11.0.0
redis>=5.0.0
celery>=5.3.0

# Logging and Monitoring
structlog>=23.0.0
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Configuration and Settings
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# Utilities
rich>=13.5.0
typer>=0.9.0
click>=8.1.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Database (optional)
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP and API
fastapi>=0.103.0
uvicorn>=0.23.0
httpx>=0.24.0

# Data Validation
marshmallow>=3.20.0
cerberus>=1.3.4

# Scheduling and Background Tasks
apscheduler>=3.10.0

# Performance and Optimization
cachetools>=5.3.0
lru-dict>=1.2.0

# File and Data Handling
openpyxl>=3.1.0
python-multipart>=0.0.6

# Demo Interface
gradio>=4.0.0

# Real-World Business Intelligence Dependencies
yfinance>=0.2.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0

# Optional Social Media API Dependencies (install if you have API keys)
# pip install tweepy praw requests-oauthlib